beta  5.0 (correct norm and change Makefile)

+++ virtual_machine/live_operation.c
@@ -33,7 +33,7 @@ int		check_who_live(unsigned int number, t_vm *main_struct)
 	i = 0;
 	while (main_struct->players[i])
 	{
-		if (number == main_struct->players[i]->name)
+		if (number == (unsigned int)main_struct->players[i]->name)
 		{
 			main_struct->players[i]->lives_in_current_period++;
 			main_struct->nbr_of_lifes++;
@@ -51,7 +51,7 @@ void	do_live_func(t_vm *main_struct, t_car *car)
 {
 	unsigned int	number;
 	int				new_life;
-	int ptr;
+	int				ptr;
 
 	car->live = 1;
 	ptr = car->pos + 1;
